infra: aws: move SDK provisioner to its own dir
This will make it easier to remove it later on when it's fully replaced
by CAPA. With the current file structure, it's hard to say whether they
share files or not.
