+++ pkg/infrastructure/aws/sdk/destroy.go
