+++ pkg/infrastructure/aws/sdk/vpc.go
